requirements.txt
requests==2.31.0
networkx>=2.6.3
pyvis>=0.2.1
matplotlib>=3.4.3
numpy==1.26.2
pandas==2.1.4
flask>=2.0.1
jinja2>=3.0.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-jwt-extended==4.6.0
flask-caching==2.1.0
flask-wtf==1.2.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-benchmark==4.0.0
pytest-timeout==2.1.0
pytest-mock==3.11.1
black==23.11.0
pylint==3.0.2
psutil==5.9.7
rdflib>=6.0.0
